Rename `ModelSourceChangedAction` etc to `SourceModelChangedAction`
This is to be consistent with the term source model that we now use to
refer to the underlying model from which the GModel is created.

https://github.com/eclipse-glsp/glsp/issues/655
