+++ plugins/org.eclipse.glsp.server/META-INF/MANIFEST.MF
@@ -26,9 +26,9 @@ Export-Package: org.eclipse.glsp.server.actions,
  org.eclipse.glsp.server.features.contextmenu,
  org.eclipse.glsp.server.features.core.model,
  org.eclipse.glsp.server.features.directediting,
- org.eclipse.glsp.server.features.modelsourcewatcher,
  org.eclipse.glsp.server.features.navigation,
  org.eclipse.glsp.server.features.popup,
+ org.eclipse.glsp.server.features.sourcemodelwatcher,
  org.eclipse.glsp.server.features.toolpalette,
  org.eclipse.glsp.server.features.undoredo,
  org.eclipse.glsp.server.features.validation,
