+++ tests/org.eclipse.glsp.server.test/src/org/eclipse/glsp/server/features/sourcemodelwatcher/FileWatcherTest.java
@@ -1,5 +1,5 @@
 /********************************************************************************
- * Copyright (c) 2020-2021 EclipseSource and others.
+ * Copyright (c) 2020-2022 EclipseSource and others.
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License v. 2.0 which is available at
@@ -13,7 +13,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  ********************************************************************************/
-package org.eclipse.glsp.server.features.modelsourcewatcher;
+package org.eclipse.glsp.server.features.sourcemodelwatcher;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
@@ -151,7 +151,7 @@ class FileWatcherTest {
    private void assertNotifications(final int size) throws InterruptedException {
       assertEquals(actionDispatcher.dispatchedActions.size(), size);
       for (int i = 0; i < size; i++) {
-         assertTrue(actionDispatcher.dispatchedActions.get(i) instanceof ModelSourceChangedAction);
+         assertTrue(actionDispatcher.dispatchedActions.get(i) instanceof SourceModelChangedAction);
       }
    }
